requirements.txt
aiogram>=3.0.0
pyrogram>=2.0.0
Tgcrypto
aiosqlite
apscheduler
google-generativeai
python-dotenv